requirements.txt
Django
python-dateutil
pyserial